Testing support and docs utility
- File access logic for /docs has been moved to its own utility in /lib/docs
- Now have the ability to run simple Jest unit tests
- package.json updated with test script and some quick shortcuts for debugging the tests
