+++ lib/posts.js
@@ -7,6 +7,7 @@ import { serialize } from 'next-mdx-remote/serialize'
 const postsDirectory = path.join(process.cwd(), 'posts');
 const docsDirectory = path.join(process.cwd(), 'docs');
 
+
 /**
  * Recursively walks the /posts directory to find all posts. Returns an array files; full file path, not just file name.
  * @param {*} dirPath 
@@ -28,6 +29,7 @@ const getAllFiles = function(dirPath, arrayOfFiles) {
 
     return arrayOfFiles;
 };
+
 
 /**
  * Returns a list of file path objects (all blog posts)
@@ -84,16 +86,6 @@ export function getAllCategories() {
     return uniqueCategories;
 };
 
-/**
- * Get a list of all sections under the /docs directory
- * @returns A string[] containing the name of each section
- */
-export function getAllDocsSections() {
-    // Read all sub-directories under the /docs directory
-
-    // Mocked an array of string(s) for now.
-    return ['owasp'];
-}
 
 /**
  * Return a list of posts containing a specific category tag
@@ -122,30 +114,7 @@ export function getCategoryPosts(category) {
     });
 }
 
-/**
- * Return posts, including frontmatter, from a specified directory.
- * @param {*} directory 
- */
-export function getPostsFromDocsSubdir(directory) {
-    const postsArr = getAllFiles(`${docsDirectory}/${directory}`);
-    console.log(`getPostsFromDocsSubdir - ${postsArr}`)
-    let docsSubdirPosts = [];
 
-    for(let i = 0; i < postsArr.length; i++) {
-        
-
-        let splits = postsArr[i].split(path.sep);
-
-        const id = splits[splits.length - 1].replace(/\.mdx$/, '');
-
-        const fileContents = fs.readFileSync(postsArr[i], 'utf8');
-        const matterResult = matter(fileContents);
-
-        docsSubdirPosts.push({ id, directory, ...matterResult.data });
-    }
-
-    return docsSubdirPosts;
-}
 
 ///
 // Fetch a specific post's data
@@ -169,27 +138,6 @@ export async function getPostData(year, id) {
     };
 }
 
-///
-// Fetch a specific doc's data
-///
-export async function getDocsData(section, id) {
-    // TODO: Support .md files
-    const fullPath = path.join(docsDirectory, `${section}/${id}.mdx`);
-    const markdownWithMeta = fs.readFileSync(fullPath, 'utf8');
-
-    // use gray-matter to parse the post metadata section
-    const frontMatter = matter(markdownWithMeta);
-
-    // frontMatter.content is the MD post
-    const mdxSource = await serialize(frontMatter.content);
-
-    // Combine the data with the id
-    return {
-        id,
-        mdxSource,
-        ...frontMatter.data
-    };
-}
 
 ///
 // Retrieve a list of file names in the `/posts` directory (excluding '.md' extension from name) 
@@ -221,43 +169,6 @@ export function getAllPostIds() {
 
     return postInfo;    
 };
-
-///
-// Retrieve a list of file names in the `/docs` directory (excluding '.md' extension from name) 
-///
-export function getAllDocsIds() {
-    let fileNames = getAllFiles(docsDirectory);
-    console.log(`getAllDocsIds - ${fileNames}`)
-    var postInfo = [];
-
-    for(var i = 0; i < fileNames.length; i++) {
-        let noExt;
-        /**
-         * Support for Markdown (.md) and MDX (.mdx) files
-         */
-        
-        // TODO : Support .md files
-        if (fileNames[i].endsWith('.mdx')) noExt = fileNames[i].replace(/\.mdx$/, '');
-        
-        console.log(`getAllDocsIds [i] - ${fileNames[i]}`)
-
-        let splitPath = noExt.split(path.sep);
-        let fileName = splitPath[splitPath.length - 1];
-        let section = splitPath[splitPath.length - 2];
-
-        console.log(`getAllDocsIds - ${fileName}, ${section}`)
-
-        postInfo.push({
-            params: {
-                id: fileName,
-                section: section
-            }
-        });
-    }
-
-    return postInfo;    
-};
-
 
 
 /**
@@ -302,37 +213,3 @@ export function getSortedPostsData() {
         }
     });
 };
-
-export function getSortedDocsPostsData() {
-    const fileNames = getAllFiles(docsDirectory); // Filenames contain the full path
-    const allPostsData = fileNames.map(fileName => {
-        // Remove the '.md' file extension, and grab the last section of the file path to...
-        // Create an id from the filename
-        var id = fileName.replace(/\.mdx$/, '');
-        id = id.split(path.sep)[id.split(path.sep).length - 1];
-
-        const year = fileName.split(path.sep)[fileName.split(path.sep).length - 2];
-        id = `${year}/${id}`;
-
-        // Read whole markdown file as string
-        const fileContents = fs.readFileSync(fileName, 'utf8');
-
-        // Use gray-matter to parse the post metadata section
-        const matterResult = matter(fileContents);
-
-        // Combine the data with the id
-        return {
-            id,
-            ...matterResult.data
-        }
-    });
-
-    // Sort posts by date
-    return allPostsData.sort((a, b) => {
-        if (a.date < b.date) {
-            return 1;
-        } else {
-            return -1;
-        }
-    });
-}
