+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
+    "test": "jest",
     "lint": "next lint",
     "debug": "cross-env NODE_OPTIONS='--inspect' next dev",
     "build": "next build",
@@ -25,8 +26,11 @@
     "react-syntax-highlighter": "^15.4.5"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
     "eslint": "7.32.0",
     "eslint-config-next": "12.0.4",
+    "jest": "^27.4.5",
     "next-mdx-remote": "^3.0.8"
   }
 }
